More fixes for clock adjustment on 16bit platforms

+++ clockless.h
@@ -9,10 +9,10 @@
 // #define NS(_NS) (_NS / (1000 / (F_CPU / 1000000L)))
 #if F_CPU < 96000000
 #define NS(_NS) ( (_NS * (F_CPU / 1000000L))) / 1000
-#define CLKS_TO_MICROS(_CLKS) _CLKS / (F_CPU / 1000000L)
+#define CLKS_TO_MICROS(_CLKS) ((long)(_CLKS)) / (F_CPU / 1000000L)
 #else
 #define NS(_NS) ( (_NS * (F_CPU / 2000000L))) / 1000
-#define CLKS_TO_MICROS(_CLKS) _CLKS / (F_CPU / 2000000L)
+#define CLKS_TO_MICROS(_CLKS) ((long)(_CLKS)) / (F_CPU / 2000000L)
 #endif
 
 //  Macro for making sure there's enough time available
